Add files via upload
Website where I will be finding most of my information:
-  https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7358007/
- May change to: https://actaneurocomms.biomedcentral.com/articles/10.1186/s40478-020-01082-6


 scientific question:
Does a mutation in the Down syndrome cell adhesion molecule-like protein 1 homolog gene (DSCAML1), cause a similar hindrance in lamina-specific synaptic connections in the retina for humans and mice?

scientific hypothesis:
If the DSCAML1 gene is mutant, then there will be retinal synaptic connection lost for humans and mice


Description of files:
I have the .html file AND the .ipynb file of my 2C (started draft of project 2). It still has a lot of work, but I plan to add a huge amount of information and coding! Everything should look better by the 2D due date (this upcoming sunday).

I included the two .fasta files that I found on NCBI for both the human and the mouse sequence for the DSCAML1 gene. I found details on that gene on this website: https://actaneurocomms.biomedcentral.com/articles/10.1186/s40478-020-01082-6